fix: un score fractionnaire, et les erreurs vont au log plutôt qu'à l'écran
Deux choses, signalées à l'usage sur le même scan.

1. finalPoints est un **flottant**. Les deux fixtures notaient un nombre entier,
   donc un int les décodait, et le premier dépôt à obtenir un score fractionnaire
   faisait échouer toute l'étape :

     cannot unmarshal number 25.698320532936123 into Go struct field
     plumberScore.finalPoints of type int

   La bannière de plumber affiche une décimale, ce qui le disait déjà. Le score
   est un flottant de bout en bout et la ligne de tête en affiche une décimale
   elle aussi — arrondir à l'entier effacerait la différence entre deux dépôts.

2. Une erreur ne s'affiche plus dans le TUI. Elle va au log, et le footer dit
   d'aller y voir (Rule 128).

   Le panneau d'avertissements remplaçait toute la table dès que Result.Errors
   n'était pas vide ; le replier en bandeau au-dessus ne faisait que déplacer le
   problème sur les cinq onglets à la fois, en permanence, pour un message lu
   une fois. Et le stderr d'un outil remis en forme pour un viewport est
   exactement ce à quoi sert un log.

   Il ne reste plus rien qui remplace la table : elle est ce que l'état
   « résultats » affiche, toujours, et le footer porte un **status** — pas un
   message — qui nomme les étapes en échec et renvoie au log. Un status parce
   que c'est un état du résultat : un message expire au bout de trois secondes,
   et l'utilisateur resterait devant un résultat qui a l'air complet.

   scan.recordStageError est ce qui rend ça honnête. Chaque étape appendait à
   Result.Errors et ne journalisait rien, donc la seule copie du motif était à
   l'écran, dans le panneau qui cachait les findings. Elle journalise maintenant
   au moment où elle enregistre, aux six endroits.

internal/ui/security/warnings.go disparaît avec ses trois fonctions de mise en
forme du stderr.

Entire-Checkpoint: ced3ca3ba1d7

+++ internal/scan/plumber.go
@@ -162,8 +162,9 @@ const (
 type PlumberReport struct {
 	// Score is the letter, empty when it was withheld.
 	Score string
-	// Points is finalPoints out of 100.
-	Points int
+	// Points is finalPoints out of 100, and it is a float: a score is a
+	// weighted sum, so a whole number is the exception rather than the shape.
+	Points float64
 	// Withheld reports a run that could not conclude — exit 3, and
 	// dataCollectionDegraded in the JSON. Its letter must never be shown: a
 	// degraded run of the fixture reads B/79 where the complete run reads E/30,
